requirements.txt
websocket-client
tqdm
requests
sanic
aiofiles
tornado